+++ impl/src/test/groovy/io/jsonwebtoken/DeprecatedJwtsTest.groovy
@@ -82,7 +82,7 @@ class DeprecatedJwtsTest {
 
     @Test
     void testHeaderWithNoArgs() {
-        def header = Jwts.header()
+        def header = Jwts.unprotectedHeader()
         assertTrue header instanceof DefaultUnprotectedHeader
     }