+++ impl/src/test/groovy/io/jsonwebtoken/impl/DefaultJwtBuilderTest.groovy
@@ -132,10 +132,10 @@ class DefaultJwtBuilderTest {
 
     @Test
     void testSetHeader() {
-        def h = Jwts.header()
+        def h = Jwts.unprotectedHeader()
         def b = new DefaultJwtBuilder()
         b.setHeader(h)
-        assertSame b.header, h
+        assertEquals h, b.header
     }
 
     @Test
@@ -267,7 +267,7 @@ class DefaultJwtBuilderTest {
     @Test
     void testCompactWithJwsHeader() {
         def b = new DefaultJwtBuilder()
-        b.setHeader(Jwts.headerBuilder().setKeyId('a'))
+        b.setHeader(Jwts.header().setKeyId('a'))
         b.setPayload('foo')
         def alg = SignatureAlgorithm.HS256
         def key = Keys.secretKeyFor(alg)
@@ -320,7 +320,7 @@ class DefaultJwtBuilderTest {
     void testSignWithKeyOnly() {
 
         def b = new DefaultJwtBuilder()
-        b.setHeader(Jwts.headerBuilder().setKeyId('a'))
+        b.setHeader(Jwts.header().setKeyId('a'))
         b.setPayload('foo')
 
         def key = KeyGenerator.getInstance('HmacSHA256').generateKey()